Filtriranje rezultata za vikendice

+++ src/main/java/com/example/demo/dto/VisitReportDTO.java
@@ -3,11 +3,11 @@ package com.example.demo.dto;
 public class VisitReportDTO {
     private Integer offerId;
     private String offerName;
-    private Integer daysVisited;
+    private Long daysVisited;
 
     public VisitReportDTO() {}
 
-    public VisitReportDTO(Integer offerId, String offerName, Integer daysVisited) {
+    public VisitReportDTO(Integer offerId, String offerName, Long daysVisited) {
         this.offerId = offerId;
         this.offerName = offerName;
         this.daysVisited = daysVisited;
@@ -29,11 +29,11 @@ public class VisitReportDTO {
         this.offerName = offerName;
     }
 
-    public Integer getDaysVisited() {
+    public long getDaysVisited() {
         return daysVisited;
     }
 
-    public void setDaysVisited(Integer daysVisited) {
+    public void setDaysVisited(Long daysVisited) {
         this.daysVisited = daysVisited;
     }
 }
